+++ src/BigCommerce/Api/ResourceApi.php
@@ -13,30 +13,17 @@ use UnexpectedValueException;
 abstract class ResourceApi extends V3ApiBase
 {
     use GetAllFromBigCommerce;
+    use GetResource;
+    use UpdateResource;
 
     abstract protected function singleResourceEndpoint(): string;
     abstract protected function multipleResourcesEndpoint(): string;
     abstract protected function resourceName(): string;
 
-    protected function getResource(): ResponseInterface
-    {
-        return $this->getClient()->getRestClient()->get($this->singleResourceUrl());
-    }
-
     protected function createResource(object $resource): ResponseInterface
     {
         return $this->getClient()->getRestClient()->post(
             $this->multipleResourceUrl(),
-            [
-                RequestOptions::JSON => $resource,
-            ]
-        );
-    }
-
-    protected function updateResource(object $resource): ResponseInterface
-    {
-        return $this->getClient()->getRestClient()->put(
-            $this->singleResourceUrl(),
             [
                 RequestOptions::JSON => $resource,
             ]