+++ RELEASE_NOTES.md
@@ -6,4 +6,4 @@ Including:
 - Customer Attributes
 - Customer Attribute Values
 - Customer Form Field Values
-
+- Customer Consent